#1674 HasColumnMeaning, ColumnMeaningCollection, DefaultColumnMeaningCollection.
