+++ api/src/main/java/org/datacleaner/metadata/HasColumnMeaning.java
@@ -1,0 +1,32 @@
+/**
+ * DataCleaner (community edition)
+ * Copyright (C) 2014 Neopost - Customer Information Management
+ *
+ * This copyrighted material is made available to anyone wishing to use, modify,
+ * copy, or redistribute it subject to the terms and conditions of the GNU
+ * Lesser General Public License, as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
+ * or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Lesser General Public License
+ * for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this distribution; if not, write to:
+ * Free Software Foundation, Inc.
+ * 51 Franklin Street, Fifth Floor
+ * Boston, MA  02110-1301  USA
+ */
+package org.datacleaner.metadata;
+
+import org.apache.metamodel.util.HasName;
+import org.datacleaner.util.HasAliases;
+
+/**
+ * A meaning of an input column used during mapping.
+ */
+public interface HasColumnMeaning extends HasName, HasAliases {
+    default String getGroup() {
+        return null;
+    }
+}
